pyproject.toml
[project]
name = "dcp"
version = "0.1.0"
description = "Docs Context Provider - Single MCP endpoint for multi-library documentation"
authors = [{ name = "DCP Contributors" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.1.1",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.14.3",
    "fastembed>=0.7.4",
    "qdrant-client>=1.17.1",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "black>=26.3.1",
    "ruff>=0.15.7",
]

[project.scripts]
dcp-server = "mcp_server.server:main"
dcp-sync = "scraper.fetcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["db", "mcp_server", "scraper"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
